feat: data layer and ui level authorization check

+++ src/features/ticket/actions/delete-ticket.ts
@@ -1,14 +1,30 @@
 "use server";
 import { setCookieByKey } from "@/actions/cookies";
-import { formErrorToActionState } from "@/components/form/utils/to-action-state";
+import {
+  formErrorToActionState,
+  toActionState,
+} from "@/components/form/utils/to-action-state";
+import { getAuthOrRedirect } from "@/features/auth/queries/get-auth-or-redirect";
+import { isOwner } from "@/features/auth/utils/is-owner";
 import { prisma } from "@/lib/prisma";
 import { ticketsPath } from "@/paths";
 import { revalidatePath } from "next/cache";
 import { redirect } from "next/navigation";
 
 export const deleteTicket = async (id: string) => {
-  await new Promise((resolve) => setTimeout(resolve, 2000));
+  const { user } = await getAuthOrRedirect(); // simulate network delay
+
   try {
+    const ticket = await prisma.ticket.findUnique({
+      where: {
+        id,
+      },
+    });
+
+    if (!ticket || !isOwner(user, ticket)) {
+      return toActionState("ERROR", "Not authorized");
+    }
+
     await prisma.ticket.delete({
       where: { id },
     });
